apply patch from Marble to build with Qt-6.9

+++ core/utilities/geolocation/engine/geodata/data/GeoDataCoordinates.cpp
@@ -642,7 +642,7 @@ QString GeoDataCoordinates::lonToString(qreal lon, GeoDataCoordinates::Notation 
             // Includes -1 case!
             if (precision < 5)
             {
-                lonString += QString::fromUtf8(" %3\"").arg(lonSec, 2, 'f', 0, QLatin1Char('0'));
+                lonString += QString::fromUtf8(" %3\"").arg((double)lonSec, 2, 'f', 0, QLatin1Char('0'));
                 return lonString + weString;
             }
 
@@ -725,7 +725,7 @@ QString GeoDataCoordinates::lonToString(qreal lon, GeoDataCoordinates::Notation 
         // Includes -1 case!
         if (precision < 5)
         {
-            lonString += QString::fromUtf8(" %3\"").arg(lonSec, 2, 'f', 0, QLatin1Char('0'));
+            lonString += QString::fromUtf8(" %3\"").arg((double)lonSec, 2, 'f', 0, QLatin1Char('0'));
             return lonString;
         }
 
@@ -862,7 +862,7 @@ QString GeoDataCoordinates::latToString(qreal lat, GeoDataCoordinates::Notation 
             // Includes -1 case!
             if (precision < 5)
             {
-                latString += QString::fromUtf8(" %3\"").arg(latSec, 2, 'f', 0, QLatin1Char('0'));
+                latString += QString::fromUtf8(" %3\"").arg((double)latSec, 2, 'f', 0, QLatin1Char('0'));
                 return latString + nsString;
             }
